1.7
*Tons of bug fixes. A LOT.
*Player power update is now once per minute,
rather than per 3 seconds.
It is also now done in async. This should
increase TPS by a lot, if you
were previously having problems with
it.
*Added /sf schedule, you can now schedule wartimes and peacetimes.
This
is seperate from "enforce relations," as it is more specific as to
what
can be changed. Still a WIP. Please report any bugs.
*Added the
option to "inject" into the chat instead of completely taking it over.
Basically adds support for essentialschat and similar plugins.