v1.84
Lots of refactoring and optimization.

I moved most of the tasks and events outside of the main .java file. I
also made all player data accessed directly from RAM, and only sometimes
saving to the disk (once per minute, and then on shutdown) for a huge
speed increase.

PS. I'm not sure what the BuildTools thing is on the last commit. That
was in fact me, I guess the github client bugged.