v1.86
World data is now also stored in memory, instead of constantly fetching
from the disk.

These past couple of commits should make simplefactions MUCH MUCH MUCH
faster when there are a lot of players online on your server, so please
update!

Also, v1.84 and up are not compatible with older versions of the plugin,
so please keep that in mind.